chore: refactor Create Wallet flow
